fix: keep the analysis-record authority inside the MCP package
The server imported the codec from the web entity tree, so the package
contract check has been red on main since the analysis archive landed:
mcp/ must reach only files it ships. The module moves into mcp/src and
the entity re-exports it, the way the receipt minting already does.
